Fix major bugs
