+++ src/server/api/routers/matchRequest.ts
@@ -9,7 +9,97 @@ import { z } from "zod";
 import { createTRPCRouter, protectedProcedure } from "~/server/api/trpc";
 import { prismaPostgres } from "~/server/db";
 import { MatchType } from "@prisma-db-psql/client";
-import { Mutex } from "~/utils/utils";
+
+
+/**
+ * A lock for synchronizing async operations.
+ * Use this to protect a critical section
+ * from getting modified by multiple async operations
+ * at the same time.
+ * https://dev.to/0916dhkim/simple-typescript-mutex-implementation-5544
+ */
+export class Mutex {
+  /**
+   * When multiple operations attempt to acquire the lock,
+   * this queue remembers the order of operations.
+   */
+  private _queue: {
+      resolve: (release: ReleaseFunction) => void;
+  }[] = [];
+
+  private _isLocked = false;
+
+  /**
+   * Wait until the lock is acquired.
+   * @returns A function that releases the acquired lock.
+   */
+  acquire() {
+      return new Promise<ReleaseFunction>((resolve) => {
+          this._queue.push({resolve});
+          this._dispatch();
+      });
+  }
+
+  /**
+   * Enqueue a function to be run serially.
+   * 
+   * This ensures no other functions will start running
+   * until `callback` finishes running.
+   * @param callback Function to be run exclusively.
+   * @returns The return value of `callback`.
+   */
+  async runExclusive<T>(callback: () => Promise<T>) {
+      const release = await this.acquire();
+      try {
+          return await callback();
+      } finally {
+          release();
+      }
+  }
+
+  /**
+   * Check the availability of the resource
+   * and provide access to the next operation in the queue.
+   *
+   * _dispatch is called whenever availability changes,
+   * such as after lock acquire request or lock release.
+   */
+  private _dispatch() {
+      if (this._isLocked) {
+          // The resource is still locked.
+          // Wait until next time.
+          return;
+      }
+      const nextEntry = this._queue.shift();
+      if (!nextEntry) {
+          // There is nothing in the queue.
+          // Do nothing until next dispatch.
+          return;
+      }
+      // The resource is available.
+      this._isLocked = true; // Lock it.
+      // and give access to the next operation
+      // in the queue.
+      nextEntry.resolve(this._buildRelease());
+  }
+
+  /**
+   * Build a release function for each operation
+   * so that it can release the lock after
+   * the operation is complete.
+   */
+  private _buildRelease(): ReleaseFunction {
+      return () => {
+          // Each release function make
+          // the resource available again
+          this._isLocked = false;
+          // and call dispatch.
+          this._dispatch();
+      };
+  }
+}
+
+type ReleaseFunction = () => void;
 
 const userObject = z.object({
   id: z.string(),
@@ -271,17 +361,8 @@ export const matchRequestRouter = createTRPCRouter({
     }),
   notifyAutomaticRequests: protectedProcedure
     .input(z.object({ requestId: z.string() }))
-    .mutation(({ ctx, input }) => {
-      ee.emit("findAutomatic", {
-        requestId: input.requestId,
-        userId: ctx.session.user.id,
-      });
-    }),
-  /**
-   * Endpoint needed to ensure that no too users are matched repeatedly
-   */
-  subscribeToAutomaticRequests: protectedProcedure.subscription(({ ctx }) => {
-    return observable<{ user1Id: string; user2Id: string }>((emit) => {
+    .mutation(async ({ ctx, input }) => {
+      
       // when a request is found, emit the two users
       async function find({
         requestId,
@@ -300,12 +381,17 @@ export const matchRequestRouter = createTRPCRouter({
         const toRunSync = async () => {
           const other = await prismaPostgres.matchRequest.findFirst({
             where: {
-              category: req.category,
-              difficulty: req.difficulty,
-              matchType: "AUTO",
-              NOT: {
-                id: requestId,
-              },
+              AND: {
+                category: req.category,
+                difficulty: req.difficulty,
+                matchType: "AUTO",
+                id: {
+                  not: requestId,
+                },
+                userId: {
+                  not: userId
+                }
+              }
             },
           });
           if (other) {
@@ -321,15 +407,26 @@ export const matchRequestRouter = createTRPCRouter({
           };
           const other = await mutex.runExclusive(toRunSync);
           if (!other) return;      
-          emit.next({
-            user1Id: ctx.session.user.id,
-            user2Id: other.id,
+          ee.emit("findAutomatic", {
+            user1Id: req.userId,
+            user2Id: other.userId,
           });
       }
+      await find({
+        requestId: input.requestId,
+        userId: ctx.session.user.id,
+      })
+    }),
+  /**
+   * Endpoint needed to ensure that no too users are matched repeatedly
+   */
+  subscribeToAutomaticRequests: protectedProcedure.subscription(({ ctx }) => {
+    return observable<{ user1Id: string; user2Id: string }>((emit) => {
       
-      const findSync = (arg1: { requestId: string; userId: string }) => {
-        void find(arg1);
+      const findSync = (arg1: { user1Id: string; user2Id: string }) => {
+        emit.next(arg1);
       };
+
       ee.on("findAutomatic", findSync);
       return () => {
         ee.off("findAutomatic", findSync);
